+++ src/iff/wav/tag/mod.rs
@@ -45,12 +45,12 @@ macro_rules! impl_accessor {
 ///
 /// * The [`TagItem`] has a value other than [`ItemValue::Binary`](crate::ItemValue::Binary)
 /// * It has a key that is 4 bytes in length and within the ASCII range
-pub struct RiffInfoList {
+pub struct RIFFInfoList {
 	/// A collection of chunk-value pairs
 	pub(crate) items: Vec<(String, String)>,
 }
 
-impl RiffInfoList {
+impl RIFFInfoList {
 	/// Get an item by key
 	pub fn get(&self, key: &str) -> Option<&str> {
 		self.items
@@ -90,7 +90,7 @@ impl RiffInfoList {
 	}
 }
 
-impl Accessor for RiffInfoList {
+impl Accessor for RIFFInfoList {
 	impl_accessor!(
 		artist  => "IART";
 		title   => "INAM";
@@ -132,7 +132,7 @@ impl Accessor for RiffInfoList {
 	}
 }
 
-impl TagExt for RiffInfoList {
+impl TagExt for RIFFInfoList {
 	type Err = LoftyError;
 
 	fn is_empty(&self) -> bool {
@@ -144,12 +144,12 @@ impl TagExt for RiffInfoList {
 	}
 
 	fn save_to(&self, file: &mut File) -> std::result::Result<(), Self::Err> {
-		RiffInfoListRef::new(self.items.iter().map(|(k, v)| (k.as_str(), v.as_str())))
+		RIFFInfoListRef::new(self.items.iter().map(|(k, v)| (k.as_str(), v.as_str())))
 			.write_to(file)
 	}
 
 	fn dump_to<W: Write>(&self, writer: &mut W) -> std::result::Result<(), Self::Err> {
-		RiffInfoListRef::new(self.items.iter().map(|(k, v)| (k.as_str(), v.as_str())))
+		RIFFInfoListRef::new(self.items.iter().map(|(k, v)| (k.as_str(), v.as_str())))
 			.dump_to(writer)
 	}
 
@@ -166,8 +166,8 @@ impl TagExt for RiffInfoList {
 	}
 }
 
-impl From<RiffInfoList> for Tag {
-	fn from(input: RiffInfoList) -> Self {
+impl From<RIFFInfoList> for Tag {
+	fn from(input: RIFFInfoList) -> Self {
 		let mut tag = Tag::new(TagType::RIFFInfo);
 
 		for (k, v) in input.items {
@@ -183,9 +183,9 @@ impl From<RiffInfoList> for Tag {
 	}
 }
 
-impl From<Tag> for RiffInfoList {
+impl From<Tag> for RIFFInfoList {
 	fn from(input: Tag) -> Self {
-		let mut riff_info = RiffInfoList::default();
+		let mut riff_info = RIFFInfoList::default();
 
 		for item in input.items {
 			if let ItemValue::Text(val) | ItemValue::Locator(val) = item.item_value {
@@ -208,19 +208,19 @@ impl From<Tag> for RiffInfoList {
 	}
 }
 
-pub(crate) struct RiffInfoListRef<'a, I>
+pub(crate) struct RIFFInfoListRef<'a, I>
 where
 	I: Iterator<Item = (&'a str, &'a str)>,
 {
 	pub(crate) items: I,
 }
 
-impl<'a, I> RiffInfoListRef<'a, I>
+impl<'a, I> RIFFInfoListRef<'a, I>
 where
 	I: Iterator<Item = (&'a str, &'a str)>,
 {
-	pub(crate) fn new(items: I) -> RiffInfoListRef<'a, I> {
-		RiffInfoListRef { items }
+	pub(crate) fn new(items: I) -> RIFFInfoListRef<'a, I> {
+		RIFFInfoListRef { items }
 	}
 
 	pub(crate) fn write_to(&mut self, file: &mut File) -> Result<()> {
@@ -254,7 +254,7 @@ pub(crate) fn tagitems_into_riff(items: &[TagItem]) -> impl Iterator<Item = (&st
 
 #[cfg(test)]
 mod tests {
-	use crate::iff::RiffInfoList;
+	use crate::iff::RIFFInfoList;
 	use crate::{Tag, TagExt, TagType};
 
 	use crate::iff::chunk::Chunks;
@@ -263,7 +263,7 @@ mod tests {
 
 	#[test]
 	fn parse_riff_info() {
-		let mut expected_tag = RiffInfoList::default();
+		let mut expected_tag = RIFFInfoList::default();
 
 		expected_tag.insert(String::from("IART"), String::from("Bar artist"));
 		expected_tag.insert(String::from("ICMT"), String::from("Qux comment"));
@@ -273,7 +273,7 @@ mod tests {
 		expected_tag.insert(String::from("IPRT"), String::from("1"));
 
 		let tag = crate::tag::utils::test_utils::read_path("tests/tags/assets/test.riff");
-		let mut parsed_tag = RiffInfoList::default();
+		let mut parsed_tag = RIFFInfoList::default();
 
 		super::read::parse_riff_info(
 			&mut Cursor::new(&tag[..]),
@@ -289,7 +289,7 @@ mod tests {
 	#[test]
 	fn riff_info_re_read() {
 		let tag = crate::tag::utils::test_utils::read_path("tests/tags/assets/test.riff");
-		let mut parsed_tag = RiffInfoList::default();
+		let mut parsed_tag = RIFFInfoList::default();
 
 		super::read::parse_riff_info(
 			&mut Cursor::new(&tag[..]),
@@ -302,7 +302,7 @@ mod tests {
 		let mut writer = Vec::new();
 		parsed_tag.dump_to(&mut writer).unwrap();
 
-		let mut temp_parsed_tag = RiffInfoList::default();
+		let mut temp_parsed_tag = RIFFInfoList::default();
 
 		// Remove the LIST....INFO from the tag
 		super::read::parse_riff_info(
@@ -321,7 +321,7 @@ mod tests {
 		let tag_bytes = crate::tag::utils::test_utils::read_path("tests/tags/assets/test.riff");
 
 		let mut reader = std::io::Cursor::new(&tag_bytes[..]);
-		let mut riff_info = RiffInfoList::default();
+		let mut riff_info = RIFFInfoList::default();
 
 		super::read::parse_riff_info(
 			&mut reader,
@@ -340,7 +340,7 @@ mod tests {
 	fn tag_to_riff_info() {
 		let tag = crate::tag::utils::test_utils::create_tag(TagType::RIFFInfo);
 
-		let riff_info: RiffInfoList = tag.into();
+		let riff_info: RIFFInfoList = tag.into();
 
 		assert_eq!(riff_info.get("INAM"), Some("Foo title"));
 		assert_eq!(riff_info.get("IART"), Some("Bar artist"));